requirements.txt
python3-rpi.gpio
python3-dev
manager
lib